@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.1.8.RELEASE</version>
+		<version>3.1.12</version>
 		<relativePath/>
 	</parent>
 	<groupId>com.slowiak</groupId>
@@ -21,12 +21,11 @@
 		<sonar.java.binaries>${project.basedir}/target</sonar.java.binaries>
 		<sonar.language>java</sonar.language>
 		<sonar.tests>${project.basedir}/target/test-classes</sonar.tests>
-		<junit-version>5.5.2</junit-version>
-		<jacoco-version>0.8.4</jacoco-version>
-		<lombok-version>1.18.10</lombok-version>
-		<jsoup-version>1.12.1</jsoup-version>
-		<java.version>1.8</java.version>
-		<spring.version>5.1.10.RELEASE</spring.version>
+		<junit-version>5.10.2</junit-version>
+		<jacoco-version>0.8.11</jacoco-version>
+		<lombok-version>1.18.30</lombok-version>
+		<jsoup-version>1.17.2</jsoup-version>
+		<java.version>17</java.version>
 	</properties>
 
 	<dependencies>
@@ -75,12 +74,6 @@
 			<artifactId>jsoup</artifactId>
 			<version>${jsoup-version}</version>
 		</dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-test</artifactId>
-            <version>${spring.version}</version>
-            <scope>compile</scope>
-        </dependency>
     </dependencies>
 
 	<build>
